etl/requirements.txt
requests>=2.31
pandas>=2.1
python-dotenv>=1.0
yfinance>=0.2.31
scipy>=1.11
